Add immutable variables (both global and local)
Local immutabes can be tracked solely in the compiler, since scopes are
restricted to a single line on the REPL.
Global immutables have to persist throughout the REPL, hence need to be
stored with the VM.
Also, prevent redefining global immutables. (Not an issue for locals,
since we prevent redefining locals in the same scope already).